better config
