+++ plugins/flytekit-flyin/flytekitplugins/flyin/__init__.py
@@ -11,9 +11,11 @@ This package contains flyin plugin for Flytekit.
    VscodeConfig
    DEFAULT_CODE_SERVER_REMOTE_PATH
    DEFAULT_CODE_SERVER_EXTENSIONS
+   COPILOT_EXTENSION
+   VIM_EXTENSION
+   CODE_TOGETHER_EXTENSION
 """
 
-from .vscode_lib.decorator import vscode, VscodeConfig
+from .vscode_lib.decorator import vscode
 from .vscode_lib.constants import DEFAULT_CODE_SERVER_REMOTE_PATH, DEFAULT_CODE_SERVER_EXTENSIONS
-from .vscode_lib.config import VIM_CONFIG_EXTENSION, COPILOT_CONFIG_EXTENSION, COPILOT_EXTENSION, VIM_EXTENSION, CODE_TOGETHER_EXTENSION
-
+from .vscode_lib.config import VscodeConfig, COPILOT_EXTENSION, VIM_EXTENSION, CODE_TOGETHER_EXTENSION
